+++ openquake/hmtk/comparison/rate_grids.py
@@ -49,8 +49,8 @@
 Python tools for calculating activity rates on a grid from a source model
 """
 import numpy as np
-from openquake.hazardlib.sourceconverter import (SourceConverter,
-                                                 area_to_point_sources)
+from openquake.hazardlib.sourceconverter import SourceConverter
+from openquake.hazardlib.source import area_to_point_sources
 from openquake.hazardlib.nrml import SourceModelParser
 from openquake.hazardlib.source.complex_fault import ComplexFaultSource
 from openquake.hazardlib.source.characteristic import CharacteristicFaultSource
