+++ openquake/qa_tests_data/classical/case_15/expected/hazard_uhs-mean-0.2.xml
@@ -18,7 +18,7 @@ xmlns:gml="http://www.opengis.net/gml"
                 </gml:pos>
             </gml:Point>
             <IMLs>
-                0.2 0.634829725052
+                0.2 0.634829675602
             </IMLs>
         </uhs>
         <uhs>
@@ -28,7 +28,7 @@ xmlns:gml="http://www.opengis.net/gml"
                 </gml:pos>
             </gml:Point>
             <IMLs>
-                0.2 0.594248574046
+                0.2 0.594248518567
             </IMLs>
         </uhs>
         <uhs>
@@ -38,7 +38,7 @@ xmlns:gml="http://www.opengis.net/gml"
                 </gml:pos>
             </gml:Point>
             <IMLs>
-                0.2 0.747483030542
+                0.2 0.747483112578
             </IMLs>
         </uhs>
     </uniformHazardSpectra>
