+++ openquake/qa_tests_data/classical/case_16/expected/quantile_curve-0.1.csv
@@ -1,3 +1,3 @@
 # quantile-0.1, investigation_time=50.0,imt=PGA
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098
-0.00000,0.00000,0.00000,9.230221E-02,9.106769E-02,8.784024E-02
+0.00000,0.00000,0.00000,9.230221E-02,9.106769E-02,8.784023E-02
