+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-PGA.csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=PGA
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098,poe-0.0137,poe-0.0192,poe-0.0269,poe-0.0376,poe-0.0527,poe-0.0738,poe-0.103,poe-0.145,poe-0.203,poe-0.284,poe-0.397,poe-0.556,poe-0.778,poe-1.09,poe-1.52,poe-2.13
-0.00000,0.00000,0.00000,9.999944E-01,9.999932E-01,9.999891E-01,9.999728E-01,9.998907E-01,9.993902E-01,9.963645E-01,9.810649E-01,9.261514E-01,7.948474E-01,5.758474E-01,3.383831E-01,1.568948E-01,5.693356E-02,1.558725E-02,3.056230E-03,3.685663E-04,5.763380E-06,0.000000E+00
+0.00000,0.00000,0.00000,9.999944E-01,9.999932E-01,9.999891E-01,9.999728E-01,9.998907E-01,9.993902E-01,9.963645E-01,9.810649E-01,9.261513E-01,7.948472E-01,5.758472E-01,3.383830E-01,1.568948E-01,5.693357E-02,1.558725E-02,3.056233E-03,3.685670E-04,5.763380E-06,0.000000E+00
