+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-PGV.csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=PGV
 lon,lat,depth,poe-2.0,poe-4.0,poe-6.0,poe-8.0,poe-10.0,poe-12.0,poe-14.0,poe-16.0,poe-18.0,poe-20.0,poe-23.0,poe-26.0,poe-29.0,poe-32.0,poe-35.0,poe-38.0,poe-41.0,poe-44.0,poe-47.0,poe-50.0,poe-55.0,poe-60.0,poe-65.0,poe-70.0,poe-75.0,poe-80.0,poe-85.0,poe-90.0,poe-95.0,poe-100.0,poe-110.0,poe-120.0,poe-130.0,poe-140.0,poe-150.0,poe-160.0,poe-170.0,poe-180.0,poe-190.0,poe-200.0,poe-220.0,poe-240.0,poe-260.0,poe-280.0,poe-300.0
-0.00000,0.00000,0.00000,9.949005E-01,9.123405E-01,7.444869E-01,5.705205E-01,4.272774E-01,3.190649E-01,2.395629E-01,1.817894E-01,1.393932E-01,1.079100E-01,7.489613E-02,5.298188E-02,3.808909E-02,2.776767E-02,2.053783E-02,1.532787E-02,1.156831E-02,8.795191E-03,6.750953E-03,5.201895E-03,3.406756E-03,2.265816E-03,1.529571E-03,1.031702E-03,7.014548E-04,4.721015E-04,3.144273E-04,2.071523E-04,1.328302E-04,8.908941E-05,3.263915E-05,8.092586E-06,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00
+0.00000,0.00000,0.00000,9.949005E-01,9.123405E-01,7.444868E-01,5.705203E-01,4.272773E-01,3.190648E-01,2.395629E-01,1.817893E-01,1.393932E-01,1.079100E-01,7.489614E-02,5.298190E-02,3.808910E-02,2.776769E-02,2.053784E-02,1.532788E-02,1.156832E-02,8.795201E-03,6.750962E-03,5.201902E-03,3.406761E-03,2.265820E-03,1.529574E-03,1.031704E-03,7.014566E-04,4.721028E-04,3.144283E-04,2.071532E-04,1.328309E-04,8.908993E-05,3.263915E-05,8.092586E-06,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00
