+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-SA(0.025).csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=SA(0.025)
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098,poe-0.0137,poe-0.0192,poe-0.0269,poe-0.0376,poe-0.0527,poe-0.0738,poe-0.103,poe-0.145,poe-0.203,poe-0.284,poe-0.397,poe-0.556,poe-0.778,poe-1.09,poe-1.52,poe-2.13
-0.00000,0.00000,0.00000,9.999945E-01,9.999935E-01,9.999901E-01,9.999774E-01,9.999166E-01,9.995606E-01,9.974132E-01,9.861666E-01,9.433187E-01,8.330847E-01,6.333562E-01,3.963397E-01,1.973769E-01,7.724355E-02,2.316328E-02,5.032441E-03,7.056785E-04,3.032725E-05,0.000000E+00
+0.00000,0.00000,0.00000,9.999945E-01,9.999935E-01,9.999901E-01,9.999774E-01,9.999166E-01,9.995606E-01,9.974132E-01,9.861665E-01,9.433187E-01,8.330846E-01,6.333560E-01,3.963396E-01,1.973769E-01,7.724355E-02,2.316329E-02,5.032444E-03,7.056796E-04,3.032752E-05,0.000000E+00
