+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-SA(0.05).csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=SA(0.05)
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098,poe-0.0137,poe-0.0192,poe-0.0269,poe-0.0376,poe-0.0527,poe-0.0738,poe-0.103,poe-0.145,poe-0.203,poe-0.284,poe-0.397,poe-0.556,poe-0.778,poe-1.09,poe-1.52,poe-2.13
-0.00000,0.00000,0.00000,9.999946E-01,9.999941E-01,9.999923E-01,9.999859E-01,9.999602E-01,9.998288E-01,9.990621E-01,9.947085E-01,9.751843E-01,9.135076E-01,7.732604E-01,5.618825E-01,3.348731E-01,1.595162E-01,5.935346E-02,1.688338E-02,3.330275E-03,3.975749E-04,4.271099E-06
+0.00000,0.00000,0.00000,9.999946E-01,9.999941E-01,9.999923E-01,9.999859E-01,9.999602E-01,9.998288E-01,9.990621E-01,9.947085E-01,9.751843E-01,9.135076E-01,7.732602E-01,5.618824E-01,3.348730E-01,1.595162E-01,5.935346E-02,1.688339E-02,3.330278E-03,3.975757E-04,4.271099E-06
