+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-SA(0.1).csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=SA(0.1)
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098,poe-0.0137,poe-0.0192,poe-0.0269,poe-0.0376,poe-0.0527,poe-0.0738,poe-0.103,poe-0.145,poe-0.203,poe-0.284,poe-0.397,poe-0.556,poe-0.778,poe-1.09,poe-1.52,poe-2.13
-0.00000,0.00000,0.00000,9.999948E-01,9.999947E-01,9.999943E-01,9.999930E-01,9.999888E-01,9.999728E-01,9.998991E-01,9.994781E-01,9.970303E-01,9.852419E-01,9.412708E-01,8.324530E-01,6.452329E-01,4.182238E-01,2.184974E-01,9.059780E-02,2.918319E-02,7.077323E-03,1.064467E-03
+0.00000,0.00000,0.00000,9.999948E-01,9.999947E-01,9.999943E-01,9.999930E-01,9.999888E-01,9.999728E-01,9.998991E-01,9.994781E-01,9.970303E-01,9.852419E-01,9.412708E-01,8.324528E-01,6.452328E-01,4.182237E-01,2.184974E-01,9.059780E-02,2.918320E-02,7.077327E-03,1.064469E-03
