+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-SA(0.2).csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=SA(0.2)
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098,poe-0.0137,poe-0.0192,poe-0.0269,poe-0.0376,poe-0.0527,poe-0.0738,poe-0.103,poe-0.145,poe-0.203,poe-0.284,poe-0.397,poe-0.556,poe-0.778,poe-1.09,poe-1.52,poe-2.13
-0.00000,0.00000,0.00000,9.999948E-01,9.999948E-01,9.999946E-01,9.999939E-01,9.999917E-01,9.999834E-01,9.999476E-01,9.997433E-01,9.984275E-01,9.908618E-01,9.564455E-01,8.561722E-01,6.661190E-01,4.287942E-01,2.228864E-01,9.391594E-02,3.188156E-02,8.671628E-03,1.733670E-03
+0.00000,0.00000,0.00000,9.999948E-01,9.999948E-01,9.999946E-01,9.999939E-01,9.999917E-01,9.999834E-01,9.999476E-01,9.997433E-01,9.984275E-01,9.908618E-01,9.564454E-01,8.561721E-01,6.661188E-01,4.287941E-01,2.228863E-01,9.391593E-02,3.188157E-02,8.671634E-03,1.733672E-03
