+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-SA(0.5).csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=SA(0.5)
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098,poe-0.0137,poe-0.0192,poe-0.0269,poe-0.0376,poe-0.0527,poe-0.0738,poe-0.103,poe-0.145,poe-0.203,poe-0.284,poe-0.397,poe-0.556,poe-0.778,poe-1.09,poe-1.52,poe-2.13
-0.00000,0.00000,0.00000,9.999945E-01,9.999935E-01,9.999903E-01,9.999788E-01,9.999261E-01,9.996252E-01,9.978017E-01,9.879768E-01,9.492692E-01,8.471959E-01,6.602978E-01,4.365543E-01,2.421568E-01,1.136357E-01,4.491345E-02,1.489173E-02,4.001205E-03,8.241378E-04,8.747549E-05
+0.00000,0.00000,0.00000,9.999945E-01,9.999935E-01,9.999903E-01,9.999788E-01,9.999261E-01,9.996251E-01,9.978017E-01,9.879768E-01,9.492692E-01,8.471958E-01,6.602977E-01,4.365542E-01,2.421568E-01,1.136357E-01,4.491346E-02,1.489174E-02,4.001210E-03,8.241394E-04,8.747595E-05
