+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-SA(1.0).csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=SA(1.0)
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098,poe-0.0137,poe-0.0192,poe-0.0269,poe-0.0376,poe-0.0527,poe-0.0738,poe-0.103,poe-0.145,poe-0.203,poe-0.284,poe-0.397,poe-0.556,poe-0.778,poe-1.09,poe-1.52,poe-2.13
-0.00000,0.00000,0.00000,9.999816E-01,9.999449E-01,9.997661E-01,9.988112E-01,9.939591E-01,9.741964E-01,9.156145E-01,7.903501E-01,6.020734E-01,3.975240E-01,2.227177E-01,1.090031E-01,4.626615E-02,1.689338E-02,5.121265E-03,1.218710E-03,1.828028E-04,2.389914E-06,0.000000E+00
+0.00000,0.00000,0.00000,9.999816E-01,9.999449E-01,9.997661E-01,9.988112E-01,9.939591E-01,9.741963E-01,9.156144E-01,7.903500E-01,6.020733E-01,3.975239E-01,2.227177E-01,1.090030E-01,4.626615E-02,1.689339E-02,5.121270E-03,1.218712E-03,1.828034E-04,2.389914E-06,0.000000E+00
