+++ openquake/qa_tests_data/classical/case_24/expected/hazard_curve-SA(2.0).csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0,imt=SA(2.0)
 lon,lat,depth,poe-0.005,poe-0.007,poe-0.0098,poe-0.0137,poe-0.0192,poe-0.0269,poe-0.0376,poe-0.0527,poe-0.0738,poe-0.103,poe-0.145,poe-0.203,poe-0.284,poe-0.397,poe-0.556,poe-0.778,poe-1.09,poe-1.52,poe-2.13
-0.00000,0.00000,0.00000,9.984597E-01,9.931895E-01,9.741936E-01,9.219305E-01,8.133869E-01,6.475393E-01,4.570768E-01,2.834354E-01,1.559676E-01,7.678980E-02,3.294259E-02,1.249357E-02,4.039057E-03,1.052007E-03,1.805176E-04,6.045813E-06,0.000000E+00,0.000000E+00,0.000000E+00
+0.00000,0.00000,0.00000,9.984597E-01,9.931895E-01,9.741936E-01,9.219304E-01,8.133868E-01,6.475391E-01,4.570766E-01,2.834353E-01,1.559675E-01,7.678980E-02,3.294260E-02,1.249358E-02,4.039061E-03,1.052009E-03,1.805182E-04,6.045813E-06,0.000000E+00,0.000000E+00,0.000000E+00
