+++ openquake/qa_tests_data/classical/case_24/expected/hazard_uhs.csv
@@ -1,3 +1,3 @@
 # source_model_tree_path=('b1',), gsim_tree_path=('b1',), investigation_time=50.0
 lon,lat,0.1~PGA,0.1~SA(0.025),0.1~SA(0.05),0.1~SA(0.1),0.1~SA(0.2),0.1~SA(0.5),0.1~SA(1.0),0.1~SA(2.0),0.02~PGA,0.02~SA(0.025),0.02~SA(0.05),0.02~SA(0.1),0.02~SA(0.2),0.02~SA(0.5),0.02~SA(1.0),0.02~SA(2.0)
-0.00000,0.00000,3.295755E-01,3.620365E-01,4.654649E-01,7.492293E-01,7.592464E-01,4.158482E-01,2.099737E-01,9.096513E-02,5.211052E-01,5.742593E-01,7.435655E-01,1.191092E+00,1.227860E+00,7.112130E-01,3.753322E-01,1.724173E-01
+0.00000,0.00000,3.295754E-01,3.620365E-01,4.654649E-01,7.492293E-01,7.592464E-01,4.158482E-01,2.099737E-01,9.096512E-02,5.211052E-01,5.742594E-01,7.435656E-01,1.191092E+00,1.227860E+00,7.112131E-01,3.753323E-01,1.724174E-01
