+++ openquake/qa_tests_data/classical_bcr/case_2/expected/bcr-structural-mean.csv
@@ -2,11 +2,11 @@ lon,lat,asset_ref,average_annual_loss_original,average_annual_loss_retrofitted,b
 8.12138E+01,2.87112E+01,a395,6.63824E+04,3.48326E+03,5.22191E+00
 8.13638E+01,2.99862E+01,a7180,1.57309E+02,1.40896E+01,3.56705E-02
 8.15138E+01,2.81862E+01,a4351,1.40217E+06,1.62180E+05,4.98120E+00
-8.24138E+01,2.89362E+01,a2925,5.06869E+04,5.84624E+03,5.18518E+00
+8.24138E+01,2.89362E+01,a2925,5.06869E+04,5.84622E+03,5.18518E+00
 8.32388E+01,2.75862E+01,a4102,9.18912E+05,1.45290E+05,5.82519E+00
-8.34638E+01,2.89362E+01,a125,1.88478E+04,1.60733E+03,6.97765E+00
-8.37638E+01,2.79612E+01,a8309,2.26306E+05,3.35428E+04,1.69140E+00
-8.39542E+01,2.82125E+01,a4498,5.25654E+07,7.89861E+06,5.78418E+00
-8.40638E+01,2.86362E+01,a5229,1.14162E+05,8.93501E+03,5.67837E+00
-8.42888E+01,2.85612E+01,a2544,1.34124E+04,2.19475E+03,6.05343E+00
-8.45138E+01,2.75112E+01,a3518,2.63436E+05,3.51515E+04,5.47510E+00
+8.34638E+01,2.89362E+01,a125,1.88496E+04,1.60788E+03,6.97815E+00
+8.37638E+01,2.79612E+01,a8309,2.26302E+05,3.35415E+04,1.69138E+00
+8.39542E+01,2.82125E+01,a4498,5.25637E+07,7.89743E+06,5.78412E+00
+8.40638E+01,2.86362E+01,a5229,1.14174E+05,8.93891E+03,5.67882E+00
+8.42888E+01,2.85612E+01,a2544,1.34161E+04,2.19636E+03,6.05455E+00
+8.45138E+01,2.75112E+01,a3518,2.63435E+05,3.51513E+04,5.47509E+00
