+++ openquake/qa_tests_data/classical_bcr/case_2/expected/bcr-structural-quantile-0.15.csv
@@ -2,11 +2,11 @@ lon,lat,asset_ref,average_annual_loss_original,average_annual_loss_retrofitted,b
 8.12138E+01,2.87112E+01,a395,6.63727E+04,3.48103E+03,5.22129E+00
 8.13638E+01,2.99862E+01,a7180,1.57292E+02,1.40845E+01,3.56674E-02
 8.15138E+01,2.81862E+01,a4351,1.40097E+06,1.61748E+05,4.97810E+00
-8.24138E+01,2.89362E+01,a2925,5.06244E+04,5.82455E+03,5.18045E+00
+8.24138E+01,2.89362E+01,a2925,5.06243E+04,5.82454E+03,5.18045E+00
 8.32388E+01,2.75862E+01,a4102,9.17849E+05,1.44054E+05,5.81866E+00
-8.34638E+01,2.89362E+01,a125,1.77360E+04,1.29796E+03,6.65287E+00
-8.37638E+01,2.79612E+01,a8309,2.16674E+05,3.02188E+04,1.63606E+00
-8.39542E+01,2.82125E+01,a4498,5.08931E+07,7.25056E+06,5.65154E+00
-8.40638E+01,2.86362E+01,a5229,1.05480E+05,6.65803E+03,5.33273E+00
-8.42888E+01,2.85612E+01,a2544,1.25347E+04,1.85381E+03,5.76375E+00
-8.45138E+01,2.75112E+01,a3518,2.59286E+05,3.34790E+04,5.41567E+00
+8.34638E+01,2.89362E+01,a125,1.77379E+04,1.29850E+03,6.65341E+00
+8.37638E+01,2.79612E+01,a8309,2.16669E+05,3.02171E+04,1.63602E+00
+8.39542E+01,2.82125E+01,a4498,5.08942E+07,7.25039E+06,5.65171E+00
+8.40638E+01,2.86362E+01,a5229,1.05490E+05,6.66055E+03,5.33316E+00
+8.42888E+01,2.85612E+01,a2544,1.25387E+04,1.85548E+03,5.76502E+00
+8.45138E+01,2.75112E+01,a3518,2.59285E+05,3.34789E+04,5.41566E+00
