+++ openquake/qa_tests_data/classical_bcr/case_2/expected/bcr-structural-quantile-0.85.csv
@@ -2,11 +2,11 @@ lon,lat,asset_ref,average_annual_loss_original,average_annual_loss_retrofitted,b
 8.12138E+01,2.87112E+01,a395,6.63883E+04,3.48460E+03,5.22229E+00
 8.13638E+01,2.99862E+01,a7180,1.57314E+02,1.40908E+01,3.56712E-02
 8.15138E+01,2.81862E+01,a4351,1.40337E+06,1.62605E+05,4.98429E+00
-8.24138E+01,2.89362E+01,a2925,5.07456E+04,5.86654E+03,5.18961E+00
-8.32388E+01,2.75862E+01,a4102,9.19612E+05,1.46422E+05,5.82780E+00
-8.34638E+01,2.89362E+01,a125,1.93180E+04,1.73849E+03,7.11485E+00
-8.37638E+01,2.79612E+01,a8309,2.30401E+05,3.49495E+04,1.71499E+00
-8.39542E+01,2.82125E+01,a4498,5.32681E+07,8.17180E+06,5.83980E+00
-8.40638E+01,2.86362E+01,a5229,1.17826E+05,9.89970E+03,5.82407E+00
-8.42888E+01,2.85612E+01,a2544,1.37833E+04,2.33909E+03,6.17566E+00
-8.45138E+01,2.75112E+01,a3518,2.65214E+05,3.58657E+04,5.50054E+00
+8.24138E+01,2.89362E+01,a2925,5.07455E+04,5.86652E+03,5.18961E+00
+8.32388E+01,2.75862E+01,a4102,9.19611E+05,1.46422E+05,5.82780E+00
+8.34638E+01,2.89362E+01,a125,1.93197E+04,1.73905E+03,7.11534E+00
+8.37638E+01,2.79612E+01,a8309,2.30397E+05,3.49483E+04,1.71497E+00
+8.39542E+01,2.82125E+01,a4498,5.32654E+07,8.17024E+06,5.83965E+00
+8.40638E+01,2.86362E+01,a5229,1.17839E+05,9.90419E+03,5.82453E+00
+8.42888E+01,2.85612E+01,a2544,1.37869E+04,2.34071E+03,6.17673E+00
+8.45138E+01,2.75112E+01,a3518,2.65214E+05,3.58655E+04,5.50053E+00
