+++ openquake/qa_tests_data/classical_tiling/case_1/expected/quantile_map-0.1.csv
@@ -1,8 +1,8 @@
 # quantile-0.1, investigation_time=50.0
 lon,lat,SA(0.1)-0.1,SA(0.1)-0.02,SA(0.2)-0.1,SA(0.2)-0.02,SA(0.5)-0.1,SA(0.5)-0.02
--0.32551,0.02562,3.791411E-01,5.970310E-01,4.242710E-01,6.816666E-01,2.303587E-01,3.928237E-01
--0.32547,-0.64888,4.732981E-01,7.847798E-01,5.136410E-01,8.583094E-01,2.807940E-01,4.924926E-01
--0.32533,-1.32337,6.137440E-02,9.615857E-02,8.057538E-02,1.246356E-01,5.797398E-02,9.008779E-02
-0.34898,0.02562,2.172353E-01,3.419138E-01,2.500337E-01,3.980609E-01,1.417185E-01,2.257793E-01
-0.34907,-0.64888,4.019419E-01,6.328949E-01,4.329427E-01,6.908287E-01,2.326170E-01,3.877667E-01
-0.34934,-1.32337,6.327427E-02,1.030457E-01,8.176702E-02,1.303583E-01,5.766465E-02,9.110130E-02
+-0.32551,0.02562,3.791402E-01,5.970331E-01,4.242708E-01,6.816692E-01,2.303582E-01,3.928253E-01
+-0.32547,-0.64888,4.732985E-01,7.847796E-01,5.136413E-01,8.583093E-01,2.807943E-01,4.924927E-01
+-0.32533,-1.32337,6.137447E-02,9.615871E-02,8.057545E-02,1.246357E-01,5.797402E-02,9.008787E-02
+0.34898,0.02562,2.172353E-01,3.419138E-01,2.500336E-01,3.980609E-01,1.417185E-01,2.257793E-01
+0.34907,-0.64888,4.019415E-01,6.328934E-01,4.329423E-01,6.908267E-01,2.326168E-01,3.877654E-01
+0.34934,-1.32337,6.327433E-02,1.030458E-01,8.176709E-02,1.303584E-01,5.766469E-02,9.110138E-02
