+++ openquake/qa_tests_data/classical_tiling/case_2/expected/hazard_map-mean.csv
@@ -12,10 +12,10 @@ lon,lat,depth,PGA-0.1,SA(0.5)-0.1
 28.23019,38.08008,0.00000,0.000000E+00,0.000000E+00
 28.26096,41.16278,0.00000,0.000000E+00,0.000000E+00
 28.64135,37.95516,0.00000,0.000000E+00,0.000000E+00
-28.84076,39.74223,-2.00000,5.784933E-02,5.573222E-02
+28.84076,39.74223,-2.00000,5.785214E-02,5.573801E-02
 28.87615,38.09587,0.00000,0.000000E+00,0.000000E+00
 28.89512,41.05510,0.00000,0.000000E+00,0.000000E+00
-29.44798,39.56320,0.00000,6.187243E-02,5.982912E-02
+29.44798,39.56320,0.00000,6.187506E-02,5.983660E-02
 29.54414,37.04174,0.00000,0.000000E+00,0.000000E+00
 29.59863,36.37907,0.00000,0.000000E+00,0.000000E+00
 29.78324,37.91932,0.00000,0.000000E+00,0.000000E+00
@@ -24,14 +24,14 @@ lon,lat,depth,PGA-0.1,SA(0.5)-0.1
 30.03091,36.43342,0.00000,0.000000E+00,0.000000E+00
 30.32822,40.28690,0.00000,0.000000E+00,0.000000E+00
 30.44782,37.94554,0.00000,0.000000E+00,0.000000E+00
-30.82767,38.68505,0.00000,5.932832E-02,7.879818E-02
+30.82767,38.68505,0.00000,5.930518E-02,7.876412E-02
 30.97827,40.91700,0.00000,0.000000E+00,0.000000E+00
 31.01540,37.09599,0.00000,0.000000E+00,0.000000E+00
 31.04469,40.75163,0.00000,0.000000E+00,0.000000E+00
 31.22851,37.71958,0.00000,0.000000E+00,0.000000E+00
 31.36420,37.80004,0.00000,0.000000E+00,0.000000E+00
 31.68568,37.68899,0.00000,0.000000E+00,0.000000E+00
-31.90401,39.01508,0.00000,4.995758E-02,5.294482E-02
+31.90401,39.01508,0.00000,4.996794E-02,5.295643E-02
 32.23257,41.12443,0.00000,0.000000E+00,0.000000E+00
 32.44178,40.46455,0.00000,0.000000E+00,0.000000E+00
 32.52082,40.96750,0.00000,0.000000E+00,0.000000E+00
