+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.02-rlz-0-PGA-10.1-40.1_Lon_Lat.csv
@@ -1,8 +1,8 @@
-# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.547691853736
+# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.547689595695
 Lon,Lat,poe
 8.25000E+00,3.97500E+01,0.00000E+00
 8.25000E+00,4.12500E+01,0.00000E+00
-9.75000E+00,3.97500E+01,1.93177E-02
+9.75000E+00,3.97500E+01,1.93180E-02
 9.75000E+00,4.12500E+01,0.00000E+00
-1.12500E+01,3.97500E+01,4.70463E-06
+1.12500E+01,3.97500E+01,4.70576E-06
 1.12500E+01,4.12500E+01,0.00000E+00
