+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.02-rlz-0-PGA-10.1-40.1_Mag.csv
@@ -1,5 +1,5 @@
-# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.547691853736
+# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.547689595695
 Mag,poe
-5.50000E+00,8.95750E-03
-6.50000E+00,7.17716E-03
-7.50000E+00,3.30507E-03
+5.50000E+00,8.95763E-03
+6.50000E+00,7.17725E-03
+7.50000E+00,3.30510E-03
