+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.02-rlz-0-PGA-10.1-40.1_Mag_Dist.csv
@@ -1,14 +1,14 @@
-# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.547691853736
+# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.547689595695
 Mag,Dist,poe
-5.50000E+00,1.25000E+01,8.95750E-03
+5.50000E+00,1.25000E+01,8.95763E-03
 5.50000E+00,3.75000E+01,0.00000E+00
 5.50000E+00,6.25000E+01,0.00000E+00
 5.50000E+00,8.75000E+01,0.00000E+00
-6.50000E+00,1.25000E+01,7.17207E-03
-6.50000E+00,3.75000E+01,5.12900E-06
+6.50000E+00,1.25000E+01,7.17215E-03
+6.50000E+00,3.75000E+01,5.13022E-06
 6.50000E+00,6.25000E+01,0.00000E+00
 6.50000E+00,8.75000E+01,0.00000E+00
-7.50000E+00,1.25000E+01,3.28442E-03
-7.50000E+00,3.75000E+01,2.07205E-05
+7.50000E+00,1.25000E+01,3.28445E-03
+7.50000E+00,3.75000E+01,2.07212E-05
 7.50000E+00,6.25000E+01,0.00000E+00
 7.50000E+00,8.75000E+01,0.00000E+00
