+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.02-rlz-0-SA(0.025)-10.1-40.1_Mag.csv
@@ -1,5 +1,5 @@
-# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.600691713864
+# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.600683836467
 Mag,poe
-5.50000E+00,8.58909E-03
-6.50000E+00,6.69115E-03
-7.50000E+00,3.09977E-03
+5.50000E+00,8.58950E-03
+6.50000E+00,6.69140E-03
+7.50000E+00,3.09987E-03
