+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.02-rlz-0-SA(0.025)-10.1-40.1_Mag_Dist.csv
@@ -1,14 +1,14 @@
-# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.600691713864
+# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.02,iml=0.600683836467
 Mag,Dist,poe
-5.50000E+00,1.25000E+01,8.58909E-03
+5.50000E+00,1.25000E+01,8.58950E-03
 5.50000E+00,3.75000E+01,0.00000E+00
 5.50000E+00,6.25000E+01,0.00000E+00
 5.50000E+00,8.75000E+01,0.00000E+00
-6.50000E+00,1.25000E+01,6.69010E-03
-6.50000E+00,3.75000E+01,1.05439E-06
+6.50000E+00,1.25000E+01,6.69035E-03
+6.50000E+00,3.75000E+01,1.05785E-06
 6.50000E+00,6.25000E+01,0.00000E+00
 6.50000E+00,8.75000E+01,0.00000E+00
-7.50000E+00,1.25000E+01,3.08173E-03
-7.50000E+00,3.75000E+01,1.80942E-05
+7.50000E+00,1.25000E+01,3.08183E-03
+7.50000E+00,3.75000E+01,1.80961E-05
 7.50000E+00,6.25000E+01,0.00000E+00
 7.50000E+00,8.75000E+01,0.00000E+00
