+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.1-rlz-0-PGA-10.1-40.1_Lon_Lat.csv
@@ -1,8 +1,8 @@
-# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.300492769469
+# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.300499775728
 Lon,Lat,poe
 8.25000E+00,3.97500E+01,0.00000E+00
 8.25000E+00,4.12500E+01,0.00000E+00
-9.75000E+00,3.97500E+01,9.28800E-02
+9.75000E+00,3.97500E+01,9.28753E-02
 9.75000E+00,4.12500E+01,0.00000E+00
-1.12500E+01,3.97500E+01,3.17618E-03
+1.12500E+01,3.97500E+01,3.17577E-03
 1.12500E+01,4.12500E+01,0.00000E+00
