+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.1-rlz-0-PGA-10.1-40.1_Mag.csv
@@ -1,5 +1,5 @@
-# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.300492769469
+# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.300499775728
 Mag,poe
-5.50000E+00,5.58113E-02
-6.50000E+00,3.22922E-02
-7.50000E+00,1.03534E-02
+5.50000E+00,5.58080E-02
+6.50000E+00,3.22905E-02
+7.50000E+00,1.03530E-02
