+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.1-rlz-0-PGA-10.1-40.1_Mag_Dist.csv
@@ -1,14 +1,14 @@
-# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.300492769469
+# smlt_path=b1,gsimlt_path=b1,imt=PGA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.300499775728
 Mag,Dist,poe
-5.50000E+00,1.25000E+01,5.55540E-02
-5.50000E+00,3.75000E+01,2.72495E-04
+5.50000E+00,1.25000E+01,5.55507E-02
+5.50000E+00,3.75000E+01,2.72421E-04
 5.50000E+00,6.25000E+01,0.00000E+00
 5.50000E+00,8.75000E+01,0.00000E+00
-6.50000E+00,1.25000E+01,2.81811E-02
-6.50000E+00,3.75000E+01,4.22019E-03
-6.50000E+00,6.25000E+01,1.02528E-05
+6.50000E+00,1.25000E+01,2.81799E-02
+6.50000E+00,3.75000E+01,4.21966E-03
+6.50000E+00,6.25000E+01,1.02483E-05
 6.50000E+00,8.75000E+01,0.00000E+00
-7.50000E+00,1.25000E+01,9.59736E-03
-7.50000E+00,3.75000E+01,6.06862E-04
-7.50000E+00,6.25000E+01,1.56593E-04
+7.50000E+00,1.25000E+01,9.59705E-03
+7.50000E+00,3.75000E+01,6.06804E-04
+7.50000E+00,6.25000E+01,1.56573E-04
 7.50000E+00,8.75000E+01,0.00000E+00
