+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.1-rlz-0-SA(0.025)-10.1-40.1_Lon_Lat.csv
@@ -1,8 +1,8 @@
-# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.3279780939
+# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.327983600317
 Lon,Lat,poe
 8.25000E+00,3.97500E+01,0.00000E+00
 8.25000E+00,4.12500E+01,0.00000E+00
-9.75000E+00,3.97500E+01,8.93677E-02
+9.75000E+00,3.97500E+01,8.93644E-02
 9.75000E+00,4.12500E+01,0.00000E+00
-1.12500E+01,3.97500E+01,2.87958E-03
+1.12500E+01,3.97500E+01,2.87931E-03
 1.12500E+01,4.12500E+01,0.00000E+00
