+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.1-rlz-0-SA(0.025)-10.1-40.1_Mag.csv
@@ -1,5 +1,5 @@
-# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.3279780939
+# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.327983600317
 Mag,poe
-5.50000E+00,5.40356E-02
-6.50000E+00,3.05422E-02
-7.50000E+00,9.88195E-03
+5.50000E+00,5.40333E-02
+6.50000E+00,3.05411E-02
+7.50000E+00,9.88168E-03
