+++ openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.1-rlz-0-SA(0.025)-10.1-40.1_Mag_Dist.csv
@@ -1,14 +1,14 @@
-# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.3279780939
+# smlt_path=b1,gsimlt_path=b1,imt=SA,investigation_time=50.0,lon=10.1,lat=40.1,poe=0.1,iml=0.327983600317
 Mag,Dist,poe
-5.50000E+00,1.25000E+01,5.38064E-02
-5.50000E+00,3.75000E+01,2.42210E-04
+5.50000E+00,1.25000E+01,5.38042E-02
+5.50000E+00,3.75000E+01,2.42162E-04
 5.50000E+00,6.25000E+01,0.00000E+00
 5.50000E+00,8.75000E+01,0.00000E+00
-6.50000E+00,1.25000E+01,2.67982E-02
-6.50000E+00,3.75000E+01,3.83995E-03
-6.50000E+00,6.25000E+01,7.20660E-06
+6.50000E+00,1.25000E+01,2.67974E-02
+6.50000E+00,3.75000E+01,3.83960E-03
+6.50000E+00,6.25000E+01,7.20387E-06
 6.50000E+00,8.75000E+01,0.00000E+00
-7.50000E+00,1.25000E+01,9.19384E-03
-7.50000E+00,3.75000E+01,5.55238E-04
-7.50000E+00,6.25000E+01,1.39329E-04
+7.50000E+00,1.25000E+01,9.19362E-03
+7.50000E+00,3.75000E+01,5.55199E-04
+7.50000E+00,6.25000E+01,1.39316E-04
 7.50000E+00,8.75000E+01,0.00000E+00
